flask_backend/requirements.txt
-i https://pypi.org/simple
absolufy-imports==0.3.1; python_full_version >= '3.6.1'
alembic==1.16.5; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.10.0; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
authlib==1.6.4; python_version >= '3.9'
bcrypt==4.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.8.3; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.3.0; python_version >= '3.10'
cloudpickle==3.1.1; python_version >= '3.8'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==46.0.1; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
dnspython==2.8.0; python_version >= '3.10'
docstring-parser==0.17.0; python_version >= '3.8'
email-validator==2.3.0; python_version >= '3.8'
fastapi==0.117.1; python_version >= '3.8'
flask==3.1.2; python_version >= '3.9'
flask-cors==6.0.1; python_version >= '3.9' and python_version < '4.0'
flask-jwt-extended==4.7.1; python_version >= '3.9' and python_version < '4'
google-adk==1.14.1; python_version >= '3.9'
google-ai-generativelanguage==0.6.15; python_version >= '3.7'
google-api-core[grpc]==2.25.1; python_version >= '3.7'
google-api-python-client==2.182.0; python_version >= '3.7'
google-auth==2.40.3; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-cloud-aiplatform[agent-engines]==1.115.0; python_version >= '3.9'
google-cloud-appengine-logging==1.6.2; python_version >= '3.7'
google-cloud-audit-log==0.3.2; python_version >= '3.7'
google-cloud-bigquery==3.38.0; python_version >= '3.9'
google-cloud-bigtable==2.32.0; python_version >= '3.7'
google-cloud-core==2.4.3; python_version >= '3.7'
google-cloud-logging==3.12.1; python_version >= '3.7'
google-cloud-resource-manager==1.14.2; python_version >= '3.7'
google-cloud-secret-manager==2.24.0; python_version >= '3.7'
google-cloud-spanner==3.58.0; python_version >= '3.9'
google-cloud-speech==2.33.0; python_version >= '3.7'
google-cloud-storage==2.19.0; python_version >= '3.7'
google-cloud-trace==1.16.2; python_version >= '3.7'
google-crc32c==1.7.1; python_version >= '3.9'
google-genai==1.38.0; python_version >= '3.9'
google-generativeai==0.8.5; python_version >= '3.9'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos[grpc]==1.70.0; python_version >= '3.7'
graphviz==0.21; python_version >= '3.9'
greenlet==3.2.4; python_version >= '3.9'
grpc-google-iam-v1==0.14.2; python_version >= '3.7'
grpc-interceptor==0.15.4; python_version >= '3.7' and python_version < '4.0'
grpcio==1.75.0; python_version >= '3.9'
grpcio-status==1.71.2; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httplib2==0.31.0; python_version >= '3.6'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.1; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.7.0; python_version >= '3.9'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jsonschema==4.25.1; python_version >= '3.9'
jsonschema-specifications==2025.9.1; python_version >= '3.9'
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==4.0.1; python_version >= '3.9'
mcp==1.14.1; python_version >= '3.10'
numpy==2.3.3; python_version >= '3.11'
opentelemetry-api==1.37.0; python_version >= '3.9'
opentelemetry-exporter-gcp-trace==1.9.0; python_version >= '3.7'
opentelemetry-resourcedetector-gcp==1.9.0a0; python_version >= '3.7'
opentelemetry-sdk==1.37.0; python_version >= '3.9'
opentelemetry-semantic-conventions==0.58b0; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==5.29.5; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.23; python_version >= '3.8'
pydantic==2.11.9; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.10.1; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
pymongo==4.15.1; python_version >= '3.9'
pyparsing==3.2.5; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.1; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pywin32==311; sys_platform == 'win32'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
rpds-py==0.27.1; python_version >= '3.9'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
shapely==2.1.1; python_version >= '3.10'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.43; python_version >= '3.7'
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3; python_version >= '3.8'
sse-starlette==3.0.2; python_version >= '3.9'
starlette==0.48.0; python_version >= '3.9'
tenacity==8.5.0; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
tzlocal==5.3.1; python_version >= '3.9'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uvicorn==0.36.0; python_version >= '3.9'
watchdog==6.0.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
zipp==3.23.0; python_version >= '3.9'
